Restoring the original file system.
