+++ INCHI-1-SRC/INCHI_EXE/inchi-1/gcc/makefile
@@ -158,7 +158,6 @@ $(P_BASE)/permutation_util.c	\
 $(P_MAIN)/dispstru.c	\
 $(P_BASE)/mol2atom.c	\
 $(P_MAIN)/ichimain.c	\
-$(P_MAIN)/inchi_main_cli.c	\
 $(P_BASE)/bcf_s.c
 #
 INCHI_OBJS = ichi_bns.o	\
@@ -207,7 +206,6 @@ sha2.o	\
 strutil.o	\
 util.o	\
 dispstru.o	\
-inchi_main_cli.o	\
 permutation_util.o	\
 ichimain.o	\
 bcf_s.o
