+++ INCHI-1-SRC/INCHI_EXE/inchi-1/vc14/inchi-1.vcxproj
@@ -69,7 +69,6 @@
     <ClCompile Include="..\..\..\INCHI_BASE\src\util.c" />
     <ClCompile Include="..\src\dispstru.c" />
     <ClCompile Include="..\src\ichimain.c" />
-    <ClCompile Include="..\src\inchi_main_cli.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\..\INCHI_BASE\src\bcf_s.h" />
